refactor: extract render functions to util
