+++ packages/panna-sdk/src/react/utils/utils.ts
@@ -1,3 +1,5 @@
+import { formatEther, formatUnits } from 'viem';
+import { TokenBalance } from '@/mocks/token-balances';
 import { lisk, liskSepolia } from '../../core';
 import { liskSepoliaTokenConfig, liskTokenConfig } from '../consts';
 import { getCountryByCode } from './countries';
@@ -82,3 +84,55 @@ export function detectUserCountry(): string | null {
     return null;
   }
 }
+
+/**
+ * Renders the fiat amount based on the token information and amount.
+ * @param tokenInfo - The token information
+ * @param amount - The amount to convert
+ * @returns The rendered fiat amount
+ */
+export const renderFiatAmount = (tokenInfo: TokenBalance, amount: string) => {
+  // This function calculates the fiat equivalent of the crypto amount
+  // using the formula: (fiatBalance * inputAmount * tokenDecimals) / tokenBalance
+  // We multiply by 10 ^ 18 to retain precision during division
+  // then format the result from wei to ether
+  return Number(
+    formatEther(
+      BigInt(
+        tokenInfo.fiatBalance.amount *
+          (!isNaN(Number(amount)) ? Number(amount) : 0) *
+          10 ** tokenInfo.token.decimals *
+          10 ** 18
+      ) / (tokenInfo.tokenBalance.value || BigInt(1))
+    )
+  ).toFixed(2);
+};
+
+/**
+ * Renders the crypto amount based on the token information and amount.
+ * @param tokenInfo - The token information
+ * @param amount - The amount to convert
+ * @returns The rendered crypto amount
+ */
+export const renderCryptoAmount = (tokenInfo: TokenBalance, amount: string) => {
+  // This function calculates the crypto equivalent of the fiat amount
+  // using the formula: (tokenBalance * inputAmount) / (fiatBalance * tokenDecimals)
+  return Number(
+    formatUnits(
+      (tokenInfo.tokenBalance.value *
+        BigInt(
+          (!isNaN(Number(amount)) ? Number(amount) : 0) *
+            10 ** tokenInfo.token.decimals
+        )) /
+        BigInt(
+          Number(
+            tokenInfo.fiatBalance.amount
+              ? Number(tokenInfo.fiatBalance.amount).toFixed(2)
+              : '1'
+          ) *
+            10 ** tokenInfo.token.decimals
+        ),
+      tokenInfo.token.decimals
+    )
+  ).toFixed(6);
+};
